First version with complete implementation
Completed the implementation of both endpoints + logic of the service

+++ src/main/java/com/mabaya/advesrtise/dto/CampaignRequestBody.java
@@ -18,7 +18,7 @@ public class CampaignRequestBody {
   @JsonProperty("start_date")
   private Long startDate;
 
-  private List<Integer> products;
+  private List<Long> products;
 
   private Double bid;
